+++ packages/express/src/middlewares/jsonBodyParser.ts
@@ -1,8 +1,7 @@
 import { json, OptionsJson } from "body-parser";
 import { NextFunction, Response } from "express";
-import { BadRequestError } from "@alanszp/errors";
-import { errorView } from "../views/errorView";
 import { GenericRequest } from "../types/GenericRequest";
+import { render400Error } from "../helpers/renderErrorJson";
 
 export function jsonBodyParser(options?: OptionsJson) {
   const bodyParser = json({ limit: "1mb", ...options });
@@ -14,10 +13,10 @@ export function jsonBodyParser(options?: OptionsJson) {
     try {
       bodyParser(req, res, (error?: unknown) => {
         if (!error) return next();
-        res.status(400).json(errorView(new BadRequestError("Malformed JSON")));
+        res.status(400).json(render400Error("Malformed JSON"));
       });
     } catch (error: unknown) {
-      res.status(400).json(errorView(new BadRequestError("Malformed JSON")));
+      res.status(400).json(render400Error("Malformed JSON"));
     }
   };
 }