+++ packages/express/src/middlewares/returnInternalServerError.ts
@@ -1,6 +1,5 @@
 import { ErrorRequestHandler, NextFunction, Response } from "express";
 import { InternalServerError } from "@alanszp/errors";
-import { errorView } from "../views/errorView";
 import { GenericRequest } from "../types/GenericRequest";
 import { ILogger } from "@alanszp/logger";
 
@@ -11,6 +10,6 @@ export type GetInternalServerErrorMiddleware = (
 export const returnInternalServerError: GetInternalServerErrorMiddleware =
   (getLogger: () => ILogger) =>
   (error: unknown, req: GenericRequest, res: Response, _next: NextFunction) => {
-    res.status(500).json(errorView(new InternalServerError(error)));
+    res.status(500).json(new InternalServerError(error).toView());
     getLogger().error("error_to_client", { error });
   };